requirements.txt
astroid==2.12.9
attrs==22.1.0
coverage==6.4.4
dill==0.3.5.1
factory-boy==3.2.1
Faker==14.2.1
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pylint==2.15.2
pyparsing==3.0.9
pytest==7.1.3
pytest-factoryboy==2.5.0
pytest-mock==3.8.2
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
tomlkit==0.11.4
typing-extensions==4.3.0
wrapt==1.14.1
